# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Fitness Tracker API",
    "version": "1.0.0",
    "description": "API for managing users, workouts, exercises, and goals. Protected routes require OAuth login."
  },
  "servers": [
    { "url": "https://your-render-url.onrender.com" }
  ],
  "paths": {
    "/users": {
      "get": {
        "summary": "Get all users",
        "responses": { "200": { "description": "List of users" } }
      },
      "post": {
        "summary": "Create a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/User" }
            }
          }
        },
        "responses": { "201": { "description": "User created" }, "400": { "description": "Validation error" } }
      }
    },
...
    "/workouts": {
      "get": {
        "summary": "Get all workouts",
        "responses": { "200": { "description": "List of workouts" } }
      },
      "post": {
        "summary": "Create a new workout",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Workout" }
            }
          }
        },
        "responses": { "201": { "description": "Workout created" }, "400": { "description": "Validation error" } }
      }
    },
...
    "/exercises": {
      "get": {
        "summary": "Get all exercises",
        "responses": { "200": { "description": "List of exercises" } }
      },
      "post": {
        "summary": "Create a new exercise",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Exercise" }
            }
          }
        },
        "responses": { "201": { "description": "Exercise created" }, "400": { "description": "Validation error" } }
      }
    },
...
    "/goals": {
      "get": {
        "summary": "Get all goals",
        "responses": { "200": { "description": "List of goals" } }
      },
      "post": {
        "summary": "Create a new goal",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Goal" }
            }
          }
        },
        "responses": { "201": { "description": "Goal created" }, "400": { "description": "Validation error" } }
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "required": ["name", "email", "age", "gender", "height", "weight", "joinedDate"],
        "properties": {
          "name": { "type": "string" },
          "email": { "type": "string" },
          "age": { "type": "integer" },
          "gender": { "type": "string" },
          "height": { "type": "number" },
          "weight": { "type": "number" },
          "joinedDate": { "type": "string", "format": "date" }
        }
      },
      "Workout": {
        "type": "object",
        "required": ["userId", "date", "type", "duration", "exercises", "notes", "caloriesBurned"],
        "properties": {
          "userId": { "type": "string" },
          "date": { "type": "string", "format": "date" },
          "type": { "type": "string" },
          "duration": { "type": "number" },
          "exercises": { "type": "array", "items": { "type": "string" } },
          "notes": { "type": "string" },
          "caloriesBurned": { "type": "number" }
        }
      },
      "Exercise": {
        "type": "object",
        "required": ["name", "muscleGroup", "equipment", "difficulty", "description", "userId", "date"],
        "properties": {
          "name": { "type": "string" },
          "muscleGroup": { "type": "string" },
          "equipment": { "type": "string" },
          "difficulty": { "type": "string" },
          "description": { "type": "string" },
          "userId": { "type": "string" },
          "date": { "type": "string", "format": "date" }
        }
      },
      "Goal": {
        "type": "object",
        "required": ["userId", "target", "type", "startDate", "endDate", "status", "notes"],
        "properties": {
          "userId": { "type": "string" },
          "target": { "type": "string" },
          "type": { "type": "string" },
          "startDate": { "type": "string", "format": "date" },
          "endDate": { "type": "string", "format": "date" },
          "status": { "type": "string" },
          "notes": { "type": "string" }
        }
      }
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid"
      }
    }
  }
}